chore: clean up the frontend package (#31)
