--- packages/frontend/app/components/layout.js
@@ -1,10 +1,0 @@
-import Nav from "./nav"
-
-const Layout = ({ children, categories, seo }) => (
-  <>
-    <Nav categories={categories} />
-    {children}
-  </>
-)
-
-export default Layout
