+++ packages/frontend/app/package.json
@@ -14,10 +14,12 @@
   },
   "dependencies": {
     "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
     "@next/env": "^12.3.0",
     "moment": "^2.24.0",
     "next": "^11.0.0",
     "qs": "^6.10.1",
+    "prop-types": "^15.8.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-markdown": "^4.2.2",
