+++ packages/frontend/app/pages/_document.jsx
@@ -1,0 +1,22 @@
+import Document, { Html, Head, Main, NextScript } from "next/document"
+
+class MyDocument extends Document {
+  render() {
+    return (
+      <Html lang="pl">
+        <Head>
+          <link
+            href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,400;0,500;0,600;0,700;1,300;1,400;1,500;1,600;1,700&display=swap"
+            rel="stylesheet"
+          />
+        </Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    )
+  }
+}
+
+export default MyDocument
